+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Fixed
 - Fixed an issue where it wasn't possible to change the site when editing an existing redirect
+- Fixed an issue where a log item's "handled" status would not persist after checking it
 
 ## 2.0.0-beta.1 - 2024-02-24
 ### Added